liblstparse: add merge lst files capability and change ctor behavior
This introduced a breaking change in how the ctor with string parameter
behave. Instead of loading the file internally, let the user send the
text file directly using a string. This is needed to make code easier
to test.

For now on, please use `LSTFile.fromFilePath("/path/to/file.lst")` instead.

Signed-off-by: Luís Ferreira <contact@lsferreira.net>